refactor(tui): extract view updater
