+++ tui/src/index.ts
@@ -4,18 +4,13 @@ import {
   Input,
   InputRenderableEvents,
   RGBA,
-  StyledText,
   Text,
-  bg,
-  bold,
   createCliRenderer,
-  fg,
   instantiate,
   type BoxRenderable,
   type CliRenderer,
   type InputRenderable,
   type KeyEvent,
-  type TextChunk,
   type TextRenderable,
 } from '@opentui/core'
 import {
@@ -23,45 +18,28 @@ import {
   getPackageDetails,
   searchPackages,
   type PackageDetails,
-  type PackageSearchResult,
 } from './search.ts'
 import {
-  compactList,
-  createField,
-  createInlineMeta,
-  formatBytes,
-  formatDate,
-  formatDownloads,
-  formatRecord,
-  isDefinedString,
-  truncateText,
-} from './app/format.ts'
-import {
   BRAILLE_SPINNER_FRAMES,
-  LIST_SCROLLBAR_WIDTH,
   SEARCH_DEBOUNCE_MS,
   SEARCH_RESULT_LIMIT,
   SPINNER_FRAME_MS,
   SPLIT_LAYOUT_MIN_WIDTH,
-  STATUS_BAR_PADDING_X,
 } from './app/constants.ts'
-import type {
-  AppState,
-  DetailStatus,
-  InspectorLine,
-  RunTuiOptions,
-  ShortcutAction,
-  StatusKind,
-} from './app/types.ts'
+import { truncateText } from './app/format.ts'
+import type { AppState, DetailStatus, RunTuiOptions } from './app/types.ts'
 import { isCtrlKey, isPlainKey, shouldQuit } from './app/keys.ts'
-import {
-  formatResultsRange,
-  hasNextResultsPage,
-  hasPreviousResultsPage,
-  selectedPackage,
-} from './app/selectors.ts'
+import { hasNextResultsPage, hasPreviousResultsPage, selectedPackage } from './app/selectors.ts'
 import { applyThemeToView, type AppThemeView } from './ui/applyTheme.ts'
+import { createViewUpdater } from './ui/viewUpdater.ts'
 import { createThemeManager, type Theme } from './theme/index.ts'
+import { createCollectionListText, createResultsFooterText } from './views/collection.ts'
+import {
+  createInspectorLines,
+  createStyledInspectorText,
+  getMaxInspectorScrollOffset,
+} from './views/inspector.ts'
+import { createShortcutBarText } from './views/shortcutBar.ts'
 
 function getRuntimeHint(error: unknown): string {
   const message = error instanceof Error ? error.message : String(error)
@@ -78,457 +56,6 @@ Current Node.js: ${process.version}`
 
 function isAbortError(error: unknown): boolean {
   return error instanceof Error && error.name === 'AbortError'
-}
-
-function createBracketSection(title: string, lines: string[]): InspectorLine[] {
-  const body = lines.filter(Boolean)
-
-  if (body.length === 0) {
-    return []
-  }
-
-  return [{ text: title, tone: 'section' }, ...body.map(line => ({ text: `  ${line}` }))]
-}
-
-function getCollectionEmptyLines(state: AppState): Array<{ title: string; detail: string }> {
-  if (!state.query.trim()) {
-    return [
-      {
-        title: 'Search npm packages',
-        detail: 'Type a package name to begin.',
-      },
-    ]
-  }
-
-  if (state.searchStatus === 'searching' && state.results.length === 0) {
-    return [
-      {
-        title: 'Searching packages...',
-        detail: `Waiting for "${truncateText(state.query, 48)}"`,
-      },
-    ]
-  }
-
-  if (state.searchStatus === 'error' && state.results.length === 0) {
-    return [
-      {
-        title: 'Failed to search packages',
-        detail: state.errorMessage ?? 'Network request failed.',
-      },
-    ]
-  }
-
-  if (state.searchStatus === 'empty') {
-    return [
-      {
-        title: `No packages found for "${truncateText(state.query, 42)}"`,
-        detail: 'Try a different package name.',
-      },
-    ]
-  }
-
-  return []
-}
-
-function getCollectionWindow(state: AppState, height: number): { start: number; end: number } {
-  const visibleRows = Math.max(1, height)
-
-  if (state.results.length <= visibleRows) {
-    return { start: 0, end: state.results.length }
-  }
-
-  const half = Math.floor(visibleRows / 2)
-  const start = Math.max(
-    0,
-    Math.min(state.selectedIndex - half, state.results.length - visibleRows),
-  )
-  return {
-    start,
-    end: Math.min(state.results.length, start + visibleRows),
-  }
-}
-
-function padCell(value: string, width: number, align: 'left' | 'right' = 'left'): string {
-  const truncated = truncateText(value, width)
-  return align === 'right' ? truncated.padStart(width, ' ') : truncated.padEnd(width, ' ')
-}
-
-function createSelectedChunk(text: string, theme: Theme): TextChunk {
-  return bg(theme.accent)(fg(theme.bg.base)(text))
-}
-
-function createCollectionListText(
-  state: AppState,
-  theme: Theme,
-  width = 80,
-  height = 12,
-): StyledText {
-  const chunks: TextChunk[] = []
-  const rowWidth = Math.max(24, width)
-  const emptyLines = getCollectionEmptyLines(state)
-
-  if (emptyLines.length > 0) {
-    emptyLines.forEach((line, index) => {
-      chunks.push(fg(theme.fg.primary)(bold(truncateText(line.title, rowWidth))))
-      chunks.push(fg(theme.fg.muted)(`\n${truncateText(line.detail, rowWidth)}`))
-      if (index < emptyLines.length - 1) {
-        chunks.push(fg(theme.fg.muted)('\n'))
-      }
-    })
-
-    return new StyledText(chunks)
-  }
-
-  const showScroll = state.results.length > height
-  const contentWidth = Math.max(22, rowWidth - (showScroll ? LIST_SCROLLBAR_WIDTH : 0))
-  const compact = contentWidth < 56
-  const versionWidth = compact ? 0 : 11
-  const downloadsWidth = compact ? 8 : 9
-  const nameWidth = compact
-    ? Math.max(8, contentWidth - 2 - downloadsWidth - 1 - 2)
-    : Math.max(16, Math.min(34, Math.floor(contentWidth * 0.34)))
-  const fixedWidth =
-    2 + nameWidth + 1 + (versionWidth > 0 ? versionWidth + 1 : 0) + downloadsWidth + 2
-  const descriptionWidth = Math.max(0, contentWidth - fixedWidth)
-  const window = getCollectionWindow(state, height)
-  const visibleCount = Math.max(1, window.end - window.start)
-  const maxIndicatorY = Math.max(0, visibleCount - 1)
-  const indicatorY = showScroll
-    ? Math.round((state.selectedIndex / Math.max(1, state.results.length - 1)) * maxIndicatorY)
-    : -1
-
-  state.results.slice(window.start, window.end).forEach((result, visibleIndex) => {
-    const actualIndex = window.start + visibleIndex
-    const selected = actualIndex === state.selectedIndex
-    const prefix = selected ? '> ' : '  '
-    const version = versionWidth > 0 ? `${padCell(`v${result.version}`, versionWidth)} ` : ''
-    const line =
-      `${prefix}${padCell(result.name, nameWidth)} ` +
-      version +
-      `${padCell(formatDownloads(result.weeklyDownloads), downloadsWidth, 'right')}  ` +
-      `${truncateText(result.description, descriptionWidth)}`
-    const paddedLine = line.padEnd(contentWidth, ' ')
-    const scrollbar = showScroll ? (visibleIndex === indicatorY ? '█' : '│') : ''
-    const isLast = visibleIndex === visibleCount - 1
-
-    if (selected) {
-      chunks.push(createSelectedChunk(paddedLine, theme))
-    } else {
-      chunks.push(fg(theme.fg.secondary)(prefix))
-      chunks.push(fg(theme.fg.primary)(bold(padCell(result.name, nameWidth))))
-      chunks.push(fg(theme.fg.muted)(' '))
-      if (versionWidth > 0) {
-        chunks.push(fg(theme.fg.secondary)(padCell(`v${result.version}`, versionWidth)))
-        chunks.push(fg(theme.fg.muted)(' '))
-      }
-      chunks.push(
-        fg(theme.fg.secondary)(
-          padCell(formatDownloads(result.weeklyDownloads), downloadsWidth, 'right'),
-        ),
-      )
-      chunks.push(
-        fg(theme.fg.muted)(
-          `  ${truncateText(result.description, descriptionWidth)}`.padEnd(
-            Math.max(0, contentWidth - fixedWidth + 2),
-            ' ',
-          ),
-        ),
-      )
-    }
-
-    if (showScroll) {
-      chunks.push(
-        fg(visibleIndex === indicatorY ? theme.accent : theme.border.subtle)(` ${scrollbar}`),
-      )
-    }
-
-    if (!isLast) {
-      chunks.push(fg(theme.fg.muted)('\n'))
-    }
-  })
-
-  return new StyledText(chunks)
-}
-
-function createInstallBlock(packageName: string): InspectorLine[] {
-  return createBracketSection('install', [
-    `npm install ${packageName}`,
-    `pnpm add ${packageName}`,
-    `yarn add ${packageName}`,
-    `bun add ${packageName}`,
-  ]).map(line =>
-    line.tone === 'section' ? line : ({ ...line, tone: 'command' } satisfies InspectorLine),
-  )
-}
-
-function createInspectorLines(
-  pkg: PackageSearchResult | undefined,
-  detail?: PackageDetails,
-  detailStatus: DetailStatus = 'idle',
-  detailError?: string,
-): InspectorLine[] {
-  if (!pkg) {
-    return [
-      { text: 'Package preview', tone: 'title' },
-      { text: '' },
-      {
-        text: 'Select a package from the collection to inspect the details available from search.',
-        tone: 'muted',
-      },
-    ]
-  }
-
-  const data = detail ?? pkg
-  const linkRows: Array<[string, string]> = []
-  if (data.links?.npm) {
-    linkRows.push(['npm', data.links.npm])
-  }
-  if (data.links?.repository) {
-    linkRows.push(['repo', data.links.repository])
-  }
-  if (data.links?.homepage) {
-    linkRows.push(['home', data.links.homepage])
-  }
-  if (data.links?.bugs) {
-    linkRows.push(['bugs', data.links.bugs])
-  }
-
-  const links = linkRows
-    .map(([label, value]) => createField(label, truncateText(value, 72)))
-    .filter(isDefinedString)
-  const keywords = compactList(data.keywords, 10)
-  const maintainers = data.maintainers
-    .map(maintainer => maintainer.username ?? maintainer.name)
-    .filter(isDefinedString)
-  const author = detail?.author?.name ?? detail?.author?.username
-  const distTags = formatRecord(detail?.distTags, 5)
-  const engineInfo = formatRecord(detail?.entryPoints?.engines, 3)
-  const binNames = detail?.entryPoints?.binNames ?? []
-  const detailStatusLine =
-    detailStatus === 'loading'
-      ? ({ text: '  Loading registry metadata...', tone: 'muted' } satisfies InspectorLine)
-      : detailStatus === 'error'
-        ? ({
-            text: `  Detail metadata unavailable: ${truncateText(detailError ?? 'request failed', 72)}`,
-            tone: 'warning',
-          } satisfies InspectorLine)
-        : undefined
-  const summaryMeta = createInlineMeta([
-    `latest v${data.version}`,
-    formatDownloads(data.weeklyDownloads),
-    data.license,
-    detail?.unpackedSize ? formatBytes(detail.unpackedSize) : undefined,
-  ])
-  const headerBlock: InspectorLine[] = [{ text: data.name, tone: 'title' }, { text: '' }]
-  if (detail?.deprecated) {
-    headerBlock.push({ text: `deprecated: ${detail.deprecated}`, tone: 'warning' })
-  }
-  if (summaryMeta) {
-    headerBlock.push({ text: summaryMeta, tone: 'muted' })
-  }
-  headerBlock.push({ text: data.description })
-  if (detailStatusLine) {
-    headerBlock.push(detailStatusLine)
-  }
-
-  const healthRows = [
-    createField('downloads', formatDownloads(data.weeklyDownloads)),
-    createField('versions', detail?.versionCount),
-    createField('maintainers', maintainers.length),
-    detail?.entryPoints?.dependenciesCount !== undefined
-      ? createField('dependencies', detail.entryPoints.dependenciesCount)
-      : undefined,
-    detail?.entryPoints?.peerDependenciesCount !== undefined
-      ? createField('peer deps', detail.entryPoints.peerDependenciesCount)
-      : undefined,
-    createField('published', detail?.date ? formatDate(detail.date) : undefined),
-    createField('created', detail?.created ? formatDate(detail.created) : undefined),
-    createField('modified', detail?.modified ? formatDate(detail.modified) : undefined),
-    createField('dist-tags', distTags),
-  ].filter(isDefinedString)
-
-  const runtimeRows = [
-    createField('type', detail?.entryPoints?.type),
-    createField('main', detail?.entryPoints?.main),
-    createField('module', detail?.entryPoints?.module),
-    createField('types', detail?.entryPoints?.types),
-    detail?.entryPoints?.hasExports !== undefined
-      ? createField('exports', detail.entryPoints.hasExports ? 'yes' : 'no')
-      : undefined,
-    createField('bin', binNames.length > 0 ? compactList(binNames, 5) : undefined),
-    createField('engines', engineInfo),
-  ].filter(isDefinedString)
-
-  const blocks: InspectorLine[][] = [
-    headerBlock,
-    createInstallBlock(data.name),
-    createBracketSection('health', healthRows),
-    createBracketSection('runtime', runtimeRows),
-    createBracketSection('links', links),
-    createBracketSection('keywords', keywords ? [keywords] : []),
-    createBracketSection(
-      'maintainers',
-      [
-        createField('author', author),
-        createField('team', maintainers.length > 0 ? compactList(maintainers, 8) : undefined),
-      ].filter(isDefinedString),
-    ),
-  ].filter(block => block.length > 0)
-
-  return blocks.flatMap((block, index) => (index === 0 ? block : [{ text: '' }, ...block]))
-}
-
-function createScrollableLines(
-  lines: InspectorLine[],
-  offset: number,
-  viewportHeight: number,
-): InspectorLine[] {
-  const visibleHeight = Math.max(1, viewportHeight)
-
-  if (lines.length <= visibleHeight) {
-    return lines
-  }
-
-  const bodyHeight = Math.max(1, visibleHeight - 1)
-  const start = Math.min(offset, Math.max(0, lines.length - bodyHeight))
-  const end = Math.min(lines.length, start + bodyHeight)
-  const indicator = `-- ${start + 1}-${end}/${lines.length} --`
-
-  return [...lines.slice(start, end), { text: indicator, tone: 'muted' }]
-}
-
-function getMaxInspectorScrollOffset(lines: InspectorLine[], viewportHeight: number): number {
-  const bodyHeight =
-    lines.length > viewportHeight ? Math.max(1, viewportHeight - 1) : viewportHeight
-
-  return Math.max(0, lines.length - bodyHeight)
-}
-
-function createStyledInspectorText(lines: InspectorLine[], theme: Theme): StyledText {
-  const chunks: TextChunk[] = []
-
-  lines.forEach((line, index) => {
-    const text = index === lines.length - 1 ? line.text : `${line.text}\n`
-
-    if (line.tone === 'title') {
-      chunks.push(fg(theme.fg.primary)(bold(text)))
-      return
-    }
-
-    if (line.tone === 'section') {
-      const newline = index === lines.length - 1 ? '' : '\n'
-      chunks.push(fg(theme.accent)(bold('[')))
-      chunks.push(fg(theme.fg.primary)(bold(line.text)))
-      chunks.push(fg(theme.accent)(bold(']')))
-      if (newline) {
-        chunks.push(fg(theme.fg.muted)(newline))
-      }
-      return
-    }
-
-    if (line.tone === 'muted') {
-      chunks.push(fg(theme.fg.muted)(text))
-      return
-    }
-
-    if (line.tone === 'command') {
-      chunks.push(fg(theme.status.success)(text))
-      return
-    }
-
-    if (line.tone === 'warning') {
-      chunks.push(fg(theme.status.warning)(text))
-      return
-    }
-
-    if (line.tone === 'danger') {
-      chunks.push(fg(theme.status.danger)(text))
-      return
-    }
-
-    chunks.push({
-      __isChunk: true,
-      text,
-    })
-  })
-
-  return new StyledText(chunks)
-}
-
-function contextActions(state: AppState): ShortcutAction[] {
-  if (state.focus === 'search') {
-    return [
-      { key: 'type', label: 'Search' },
-      { key: 'enter', label: 'Results' },
-      { key: '↑/↓', label: 'Results' },
-      { key: 'esc', label: 'Cancel' },
-    ]
-  }
-
-  if (state.layout === 'single' && state.view === 'inspector') {
-    return [
-      { key: 'h/esc', label: 'Results' },
-      { key: 'j/k', label: 'Scroll' },
-      { key: '/', label: 'Search' },
-      { key: 'q', label: 'Quit' },
-    ]
-  }
-
-  if (state.focus === 'inspector') {
-    return [
-      { key: 'h', label: 'Results' },
-      { key: 'j/k', label: 'Scroll' },
-      { key: '/', label: 'Search' },
-      { key: 'q', label: 'Quit' },
-    ]
-  }
-
-  return [
-    { key: 'j/k', label: 'Navigate' },
-    { key: '[/]', label: 'Page' },
-    { key: 'l', label: 'Details' },
-    { key: 'enter', label: 'Preview' },
-    { key: '/', label: 'Search' },
-    { key: 'q', label: 'Quit' },
-  ]
-}
-
-function createShortcutBarText(state: AppState, theme: Theme, width = 0): StyledText {
-  const chunks: TextChunk[] = []
-  const actions = contextActions(state)
-  const brand = './npmx'
-  const padding = ' '.repeat(STATUS_BAR_PADDING_X)
-  const availableWidth = Math.max(0, width - STATUS_BAR_PADDING_X * 2)
-  const shortcutsLength = actions.reduce(
-    (length, action, index) =>
-      length + (index > 0 ? 3 : 0) + action.key.length + 1 + action.label.length,
-    0,
-  )
-
-  chunks.push(fg(theme.fg.muted)(padding))
-
-  actions.forEach((action, index) => {
-    if (index > 0) {
-      chunks.push(fg(theme.fg.muted)('   '))
-    }
-
-    chunks.push(fg(theme.accent)(bold(action.key)))
-    chunks.push(fg(theme.fg.secondary)(` ${action.label}`))
-  })
-
-  const spacerWidth = availableWidth - shortcutsLength - brand.length
-  chunks.push(fg(theme.fg.muted)(spacerWidth > 0 ? ' '.repeat(spacerWidth) : '   '))
-  chunks.push(fg(theme.fg.primary)(bold(brand)))
-  chunks.push(fg(theme.fg.muted)(padding))
-
-  return new StyledText(chunks)
-}
-
-function createResultsFooterText(state: AppState, theme: Theme): StyledText {
-  const complete = state.query.trim().length > 0 && state.total > 0 && !hasNextResultsPage(state)
-  const rangeColor = complete ? theme.fg.primary : theme.fg.muted
-
-  return new StyledText([fg(theme.fg.muted)('\n'), fg(rangeColor)(formatResultsRange(state))])
 }
 
 export async function runTui(options: RunTuiOptions = {}): Promise<void> {
@@ -812,73 +339,17 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
     statusBar,
   }
 
-  function getStatusBarWidth(): number {
-    return Math.max(1, Number(statusBar.width) || renderer.terminalWidth)
-  }
-
-  function setStatus(message: string, kind: StatusKind = 'info'): void {
-    state.statusMessage = message
-    state.statusKind = kind
-    statusBar.content = createShortcutBarText(state, theme, getStatusBarWidth())
-    statusBar.fg = theme.status[kind]
-  }
-
-  function updateStatusBar(): void {
-    statusBar.content = createShortcutBarText(state, theme, getStatusBarWidth())
-    statusBar.fg = theme.status[state.statusKind]
-  }
-
-  function updateCollectionTitle(): void {
-    collectionPane.title = ' Results '
-    searchPanel.bottomTitle = state.query.trim()
-      ? `${state.results.length}/${state.total}`
-      : 'npm registry'
-    resultsFooter.content = createResultsFooterText(state, theme)
-  }
-
-  function updateFocusStyles(): void {
-    searchPanel.titleColor = state.focus === 'search' ? theme.accent : theme.fg.secondary
-    searchPanel.borderColor = state.focus === 'search' ? theme.accent : theme.border.normal
-    collectionPane.titleColor = state.focus === 'collection' ? theme.accent : theme.fg.secondary
-    collectionPane.borderColor = state.focus === 'collection' ? theme.accent : theme.border.normal
-    inspectorPane.titleColor = state.focus === 'inspector' ? theme.accent : theme.fg.secondary
-    inspectorPane.borderColor = state.focus === 'inspector' ? theme.accent : theme.border.normal
-
-    updateStatusBar()
-  }
-
-  function updateInspector(): void {
-    const pkg = selectedPackage(state)
-    const detail = pkg ? detailCache.get(pkg.name) : undefined
-    const content = createInspectorLines(pkg, detail, detailStatus, detailError)
-    const viewportHeight = Math.max(1, inspector.height || 1)
-
-    state.inspectorScrollOffset = Math.min(
-      state.inspectorScrollOffset,
-      getMaxInspectorScrollOffset(content, viewportHeight),
-    )
-
-    inspectorPane.title = ' Preview '
-    inspector.content = createStyledInspectorText(
-      createScrollableLines(content, state.inspectorScrollOffset, viewportHeight),
-      theme,
-    )
-    updateFocusStyles()
-  }
-
-  function updateCollection(): void {
-    const collectionWidth = Math.max(1, Number(collectionList.width) || 80)
-    const collectionHeight = Math.max(1, Number(collectionList.height) || 12)
-    state.selectedIndex = Math.min(state.selectedIndex, Math.max(0, state.results.length - 1))
-    collectionList.content = createCollectionListText(
-      state,
-      theme,
-      collectionWidth,
-      collectionHeight,
-    )
-    updateCollectionTitle()
-    updateInspector()
-  }
+  const viewUpdater = createViewUpdater({
+    renderer,
+    view: appThemeView,
+    state,
+    getTheme: () => theme,
+    getDetailState: () => ({
+      detailStatus,
+      detailError,
+      detailCache,
+    }),
+  })
 
   async function loadSelectedPackageDetails(): Promise<void> {
     const pkg = selectedPackage(state)
@@ -888,14 +359,14 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
     if (!pkg) {
       detailStatus = 'idle'
       detailError = undefined
-      updateInspector()
+      viewUpdater.updateInspector()
       return
     }
 
     if (detailCache.has(pkg.name)) {
       detailStatus = 'success'
       detailError = undefined
-      updateInspector()
+      viewUpdater.updateInspector()
       return
     }
 
@@ -904,7 +375,7 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
     activeDetailRequest = controller
     detailStatus = 'loading'
     detailError = undefined
-    updateInspector()
+    viewUpdater.updateInspector()
 
     try {
       const detail = await getPackageDetails({
@@ -920,7 +391,7 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
       detailCache.set(pkg.name, detail)
       detailStatus = 'success'
       detailError = undefined
-      updateInspector()
+      viewUpdater.updateInspector()
     } catch (error) {
       if (
         controller.signal.aborted ||
@@ -932,35 +403,8 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
 
       detailStatus = 'error'
       detailError = error instanceof Error ? error.message : String(error)
-      updateInspector()
+      viewUpdater.updateInspector()
     }
-  }
-
-  function applyLayout(): void {
-    state.layout = renderer.terminalWidth >= SPLIT_LAYOUT_MIN_WIDTH ? 'split' : 'single'
-
-    if (state.layout === 'split') {
-      state.view = 'collection'
-      workspace.flexDirection = 'row'
-      leftPane.visible = true
-      inspectorPane.visible = true
-      leftPane.width = '43%'
-      inspectorPane.width = '57%'
-      leftPane.flexGrow = 0
-      inspectorPane.flexGrow = 0
-    } else {
-      workspace.flexDirection = 'column'
-      const showingInspector = state.view === 'inspector'
-      leftPane.visible = !showingInspector
-      inspectorPane.visible = showingInspector
-      leftPane.width = '100%'
-      inspectorPane.width = '100%'
-      leftPane.flexGrow = showingInspector ? 0 : 1
-      inspectorPane.flexGrow = showingInspector ? 1 : 0
-    }
-
-    updateCollection()
-    updateStatusBar()
   }
 
   function stopSpinner(): void {
@@ -991,8 +435,8 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
     }
     input.showCursor = true
     input.focus()
-    setStatus('Search focused')
-    applyLayout()
+    viewUpdater.setStatus('Search focused')
+    viewUpdater.applyLayout()
   }
 
   function focusCollection(): void {
@@ -1003,12 +447,12 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
 
     if (state.layout === 'single') {
       state.view = 'collection'
-      applyLayout()
+      viewUpdater.applyLayout()
     } else {
-      updateFocusStyles()
+      viewUpdater.updateFocusStyles()
     }
 
-    setStatus('Results focused')
+    viewUpdater.setStatus('Results focused')
   }
 
   function focusInspector(): boolean {
@@ -1024,12 +468,12 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
 
     if (state.layout === 'single') {
       state.view = 'inspector'
-      applyLayout()
+      viewUpdater.applyLayout()
     } else {
-      updateFocusStyles()
+      viewUpdater.updateFocusStyles()
     }
 
-    setStatus(`Details focused: ${pkg.name}@${pkg.version}`)
+    viewUpdater.setStatus(`Details focused: ${pkg.name}@${pkg.version}`)
     return true
   }
 
@@ -1045,11 +489,11 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
     }
 
     state.inspectorScrollOffset = 0
-    updateCollection()
+    viewUpdater.updateCollection()
     void loadSelectedPackageDetails()
     const pkg = selectedPackage(state)
     if (pkg) {
-      setStatus(`${pkg.name}@${pkg.version}`)
+      viewUpdater.setStatus(`${pkg.name}@${pkg.version}`)
     }
   }
 
@@ -1069,13 +513,13 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
         : Math.min(maxOffset, state.inspectorScrollOffset + amount)
 
     if (nextOffset === state.inspectorScrollOffset) {
-      setStatus(direction === 'up' ? 'Top of details' : 'End of details')
+      viewUpdater.setStatus(direction === 'up' ? 'Top of details' : 'End of details')
       return
     }
 
     state.inspectorScrollOffset = nextOffset
-    updateInspector()
-    setStatus(`Details ${state.inspectorScrollOffset + 1}/${maxOffset + 1}`)
+    viewUpdater.updateInspector()
+    viewUpdater.setStatus(`Details ${state.inspectorScrollOffset + 1}/${maxOffset + 1}`)
   }
 
   function openSelection(): void {
@@ -1086,7 +530,7 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
 
     state.inspectorScrollOffset = 0
     focusInspector()
-    setStatus(`Previewing ${pkg.name}@${pkg.version}`)
+    viewUpdater.setStatus(`Previewing ${pkg.name}@${pkg.version}`)
   }
 
   function showCollection(): boolean {
@@ -1100,7 +544,7 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
 
   function focusResultsFromSearch(position: 'current' | 'first' | 'last' = 'current'): boolean {
     if (state.results.length === 0) {
-      setStatus('No results to focus')
+      viewUpdater.setStatus('No results to focus')
       return true
     }
 
@@ -1112,7 +556,7 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
 
     state.inspectorScrollOffset = 0
     focusCollection()
-    updateCollection()
+    viewUpdater.updateCollection()
     void loadSelectedPackageDetails()
     return true
   }
@@ -1132,12 +576,12 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
         : state.pageOffset + SEARCH_RESULT_LIMIT
 
     if (direction === 'previous' && !hasPreviousResultsPage(state)) {
-      setStatus('First results page')
+      viewUpdater.setStatus('First results page')
       return true
     }
 
     if (direction === 'next' && !hasNextResultsPage(state)) {
-      setStatus('Last results page')
+      viewUpdater.setStatus('Last results page')
       return true
     }
 
@@ -1166,8 +610,8 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
       state.selectedIndex = 0
       state.inspectorScrollOffset = 0
       state.errorMessage = undefined
-      updateCollection()
-      setStatus(state.focus === 'search' ? 'Search focused' : 'Results focused')
+      viewUpdater.updateCollection()
+      viewUpdater.setStatus(state.focus === 'search' ? 'Search focused' : 'Results focused')
       return
     }
 
@@ -1176,8 +620,8 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
     state.searchStatus = 'searching'
     state.errorMessage = undefined
     startSpinner()
-    updateCollection()
-    setStatus(`Searching "${truncateText(trimmed, 48)}"`)
+    viewUpdater.updateCollection()
+    viewUpdater.setStatus(`Searching "${truncateText(trimmed, 48)}"`)
 
     try {
       const response = await searchPackages({
@@ -1200,9 +644,9 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
       state.inspectorScrollOffset = 0
       state.searchStatus = response.results.length > 0 ? 'success' : 'empty'
       state.errorMessage = undefined
-      updateCollection()
+      viewUpdater.updateCollection()
       void loadSelectedPackageDetails()
-      setStatus(
+      viewUpdater.setStatus(
         response.results.length > 0
           ? `${response.results.length} packages found`
           : `No packages found for "${truncateText(trimmed, 48)}"`,
@@ -1216,8 +660,8 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
       stopSpinner()
       state.searchStatus = 'error'
       state.errorMessage = error instanceof Error ? error.message : String(error)
-      updateCollection()
-      setStatus('Package search failed', 'danger')
+      viewUpdater.updateCollection()
+      viewUpdater.setStatus('Package search failed', 'danger')
     }
   }
 
@@ -1237,8 +681,8 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
 
     stopSpinner()
     state.searchStatus = 'debouncing'
-    updateCollectionTitle()
-    updateStatusBar()
+    viewUpdater.updateCollectionTitle()
+    viewUpdater.updateStatusBar()
 
     debounceTimer = setTimeout(() => {
       debounceTimer = undefined
@@ -1359,7 +803,7 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
 
   renderer.keyInput.on('keypress', quitHandler)
   renderer.keyInput.on('keypress', modeHandler)
-  renderer.on(CliRenderEvents.RESIZE, applyLayout)
+  renderer.on(CliRenderEvents.RESIZE, viewUpdater.applyLayout)
 
   function applyTheme(nextTheme: Theme): void {
     theme = nextTheme
@@ -1371,7 +815,7 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
       theme,
       createCollectionListText,
       createResultsFooterText,
-      updateInspector,
+      updateInspector: viewUpdater.updateInspector,
     })
   }
 
@@ -1387,13 +831,13 @@ export async function runTui(options: RunTuiOptions = {}): Promise<void> {
     input.off(InputRenderableEvents.INPUT, scheduleSearch)
     renderer.keyInput.off('keypress', modeHandler)
     renderer.keyInput.off('keypress', quitHandler)
-    renderer.off(CliRenderEvents.RESIZE, applyLayout)
+    renderer.off(CliRenderEvents.RESIZE, viewUpdater.applyLayout)
     themeManager.dispose()
   })
 
   renderer.root.add(shell)
-  applyLayout()
-  updateCollection()
+  viewUpdater.applyLayout()
+  viewUpdater.updateCollection()
   focusSearch()
 }
